refactor(ENTESB-12218): Start to refactor api-generator for OpenAPI 3.x support
