+++ app/server/api-generator/src/main/java/io/syndesis/server/api/generator/openapi/util/OpenApiModelParser.java
@@ -14,34 +14,29 @@
  * limitations under the License.
  */
 
-package io.syndesis.server.api.generator.swagger.util;
+package io.syndesis.server.api.generator.openapi.util;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Locale;
-import java.util.Optional;
 import java.util.regex.Pattern;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.github.fge.jsonschema.core.exceptions.ProcessingException;
-import com.github.fge.jsonschema.core.load.configuration.LoadingConfiguration;
-import com.github.fge.jsonschema.core.report.ProcessingMessage;
-import com.github.fge.jsonschema.core.report.ProcessingReport;
-import com.github.fge.jsonschema.main.JsonSchema;
-import com.github.fge.jsonschema.main.JsonSchemaFactory;
 import io.apicurio.datamodels.Library;
 import io.apicurio.datamodels.core.models.Document;
+import io.apicurio.datamodels.openapi.models.OasDocument;
 import io.apicurio.datamodels.openapi.v2.models.Oas20Document;
+import io.apicurio.datamodels.openapi.v3.models.Oas30Document;
 import io.syndesis.common.model.Violation;
 import io.syndesis.common.util.IOStreams;
-import io.syndesis.common.util.Resources;
 import io.syndesis.common.util.json.JsonUtils;
 import io.syndesis.server.api.generator.APIValidationContext;
-import io.syndesis.server.api.generator.swagger.OpenApiModelInfo;
-import io.syndesis.server.api.generator.swagger.OpenApiValidationRules;
+import io.syndesis.server.api.generator.openapi.OpenApiModelInfo;
+import io.syndesis.server.api.generator.openapi.OpenApiSchemaValidator;
+import io.syndesis.server.api.generator.openapi.v2.Oas20SchemaValidator;
+import io.syndesis.server.api.generator.openapi.v3.Oas30SchemaValidator;
+import io.syndesis.server.api.generator.openapi.OpenApiValidationRules;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,33 +47,15 @@ import static java.util.Optional.ofNullable;
 /**
  * @author Christoph Deppisch
  */
-public final class Oas20ModelParser {
+public final class OpenApiModelParser {
 
     private static final Pattern JSON_TEST = Pattern.compile("^\\s*\\{.*");
 
-    private static final Logger LOG = LoggerFactory.getLogger(Oas20ModelParser.class);
-
-    private static final JsonSchema SWAGGER_2_0_SCHEMA;
-
-    private static final String SWAGGER_2_0_SCHEMA_FILE = "schema/swagger-2.0-schema.json";
-
-    private static final String SWAGGER_IO_V2_SCHEMA_URI = "http://swagger.io/v2/schema.json#";
+    private static final Logger LOG = LoggerFactory.getLogger(OpenApiModelParser.class);
 
     private static final Yaml YAML_PARSER = new Yaml();
 
-    static {
-        try {
-            final JsonNode swagger20Schema = JsonUtils.reader().readTree(Resources.getResourceAsText(SWAGGER_2_0_SCHEMA_FILE));
-            final LoadingConfiguration loadingConfiguration = LoadingConfiguration.newBuilder()
-                .preloadSchema(SWAGGER_IO_V2_SCHEMA_URI, swagger20Schema)
-                .freeze();
-            SWAGGER_2_0_SCHEMA = JsonSchemaFactory.newBuilder().setLoadingConfiguration(loadingConfiguration).freeze().getJsonSchema(SWAGGER_IO_V2_SCHEMA_URI);
-        } catch (final ProcessingException | IOException ex) {
-            throw new IllegalStateException("Unable to load the schema file embedded in the artifact", ex);
-        }
-    }
-
-    private Oas20ModelParser() {
+    private OpenApiModelParser() {
         // utility class
     }
 
@@ -122,23 +99,51 @@ public final class Oas20ModelParser {
         }
 
         final Document parsed = Library.readDocumentFromJSONString(resolvedSpecification);
-        if (!(parsed instanceof Oas20Document)) {
+        if (!(parsed instanceof OasDocument)) {
             LOG.debug("Unable to read OpenAPI document\n{}\n", specification);
             return resultBuilder
-                .addError(new Violation.Builder().error("error").property("").message("Unable to read OpenAPI document from: '"
-                    + StringUtils.abbreviate(specification, 100)).build())
+                .addError(new Violation.Builder()
+                    .error("error")
+                    .property("")
+                    .message("Unable to read OpenAPI document from: '" + StringUtils.abbreviate(resolvedSpecification, 100)).build())
                 .build();
         }
 
+        resultBuilder.model((OasDocument) parsed);
         if (validationContext != APIValidationContext.NONE) {
-            final OpenApiModelInfo swaggerModelInfo = validateJSonSchema(resolvedSpecification, (Oas20Document) parsed);
-            return OpenApiValidationRules.get(validationContext).apply(swaggerModelInfo);
+            try {
+                getSchemaValidator((OasDocument) parsed).validateJSonSchema(convertToJson(resolvedSpecification), resultBuilder);
+            } catch (IOException e) {
+                return resultBuilder
+                    .addError(new Violation.Builder()
+                        .error("error")
+                        .property("")
+                        .message("Unable to read OpenAPI document from: '" + StringUtils.abbreviate(resolvedSpecification, 100)).build())
+                    .build();
+            }
+            return OpenApiValidationRules.get(validationContext).apply(resultBuilder.build());
         }
-        return resultBuilder.model((Oas20Document) parsed).build();
+        return resultBuilder.build();
+    }
+
+    public static boolean isJsonSpec(final String specification) {
+        return JSON_TEST.matcher(specification).matches();
+    }
+
+    private static OpenApiSchemaValidator getSchemaValidator(OasDocument openApiDoc) {
+        if (openApiDoc instanceof Oas20Document) {
+            return new Oas20SchemaValidator();
+        }
+
+        if (openApiDoc instanceof Oas30Document) {
+            return new Oas30SchemaValidator();
+        }
+
+        throw new IllegalStateException(String.format("Unable to determine proper schema validator for OpenAPI document type '%s'", openApiDoc.getClass()));
     }
 
     @SuppressWarnings("PMD.SignatureDeclareThrowsException")
-    static String resolve(final String specification) throws Exception {
+    private static String resolve(final String specification) throws Exception {
         final String specificationToUse;
         if (specification.toLowerCase(Locale.US).startsWith("http")) {
             specificationToUse = resolve(new URL(specification));
@@ -151,7 +156,7 @@ public final class Oas20ModelParser {
         return JsonUtils.writer().writeValueAsString(node);
     }
 
-    static String resolve(URL url) {
+    private static String resolve(URL url) {
         HttpURLConnection con = null;
         try {
             con = (HttpURLConnection) url.openConnection();
@@ -173,53 +178,9 @@ public final class Oas20ModelParser {
         }
     }
 
-    private static OpenApiModelInfo validateJSonSchema(final String specification, final Oas20Document model) {
-        try {
-            final JsonNode specRoot = convertToJson(specification);
-            final ProcessingReport report = SWAGGER_2_0_SCHEMA.validate(specRoot);
-            final List<Violation> errors = new ArrayList<>();
-            final List<Violation> warnings = new ArrayList<>();
-            for (final ProcessingMessage message : report) {
-                final boolean added = append(errors, message, Optional.of("error"));
-                if (!added) {
-                    append(warnings, message, Optional.empty());
-                }
-            }
-
-            return new OpenApiModelInfo.Builder().errors(errors).warnings(warnings).model(model).resolvedSpecification(specification)
-                .build();
-
-        } catch (IOException | ProcessingException ex) {
-            LOG.error("Unable to load the schema file embedded in the artifact", ex);
-            return new OpenApiModelInfo.Builder().addError(new Violation.Builder().error("error").property("")
-                .message("Unable to load the OpenAPI schema file embedded in the artifact").build()).build();
-        }
-    }
-
-    private static boolean append(final List<Violation> violations, final ProcessingMessage message, final Optional<String> requiredLevel) {
-        if (requiredLevel.isPresent()) {
-            final Optional<String> level = ofNullable(message.asJson()).flatMap(node -> ofNullable(node.get("level")))
-                .flatMap(node -> ofNullable(node.textValue()));
-
-            if (!level.equals(requiredLevel)) {
-                return false; // skip
-            }
-        }
-
-        final Optional<String> property = ofNullable(message.asJson()).flatMap(node -> ofNullable(node.get("instance")))
-            .flatMap(node -> ofNullable(node.get("pointer"))).flatMap(node -> ofNullable(node.textValue()));
-
-        final Optional<String> error = ofNullable(message.asJson()).flatMap(node -> ofNullable(node.get("domain")))
-            .flatMap(node -> ofNullable(node.textValue()));
-
-        violations.add(new Violation.Builder().error(error.orElse("")).message(message.getMessage()).property(property.orElse("")).build());
-
-        return true;
-    }
-
     static JsonNode convertToJson(final String specification) throws IOException {
         final JsonNode specRoot;
-        if (JSON_TEST.matcher(specification).matches()) {
+        if (isJsonSpec(specification)) {
             specRoot = JsonUtils.reader().readTree(specification);
         } else {
             specRoot = JsonUtils.convertValue(YAML_PARSER.load(specification), JsonNode.class);
