+++ app/server/api-generator/src/main/java/io/syndesis/server/api/generator/openapi/v2/CyclicValidationCheck.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.server.api.generator.swagger;
+package io.syndesis.server.api.generator.openapi.v2;
 
 import java.util.HashSet;
 import java.util.List;
