+++ app/server/api-generator/src/main/java/io/syndesis/server/api/generator/swagger/UnifiedDataShapeGenerator.java
@@ -21,6 +21,7 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.apicurio.datamodels.openapi.v2.models.Oas20Document;
 import io.apicurio.datamodels.openapi.v2.models.Oas20Operation;
 import io.syndesis.common.model.DataShape;
+import io.syndesis.server.api.generator.openapi.DataShapeGenerator;
 
 public final class UnifiedDataShapeGenerator implements DataShapeGenerator {
