+++ app/server/api-generator/src/test/java/io/syndesis/server/api/generator/openapi/v2/CyclicValidationCheckTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.server.api.generator.swagger;
+package io.syndesis.server.api.generator.openapi.v2;
 
 import java.io.IOException;
 import java.util.ArrayList;
